process_data/requirements.txt
newspaper3k
redis
pymongo
faiss-cpu
langchain
pydantic
sentence-transformers
lxml_html_clean
google-generativeai
flask